Final changes to readme.md instructions + favico
Ran through the installation again and made a number of changes to the
instructions.  They should be pretty set now.  Hopefully you can escape
angular brackets in markdown with a backslash or I'll be committing a
change for that shortly.  :)

I also added a favico.ico to the public directory.