Add and delete/disable contexts
On the Add Tasks page, there is now a + for adding new contexts and on
contexts that are not system contexts there is a - for deleting a
context.  Deleting a context that already has projects associated with
it will disable (soft delete) the context so foreign key integrity is
not violated.  But contexts with no associated projects will be deleted
from the DB.

Currently it seems as though the ability to filter contexts in the
dashboard by clicking on the context label next to a given task is
broken.  We can leave this for now and fix it under a different
feature branch.