Force Auth::id() to be integer for strict checks
For some reason in the Dev implementation of Projectify Auth::id()
returns the user's ID as a string.  This is not the case in Production.
Rather than troubleshoot Laravel 4.2 to get Auth::id() to return
integers, we'll just cast the result to int every time we call it for
now.